logparsing/Lightweight_LLM/requirements.txt
transformers==4.45.2 
pymongo
torch
python-dotenv
pandas